requirements.txt
alembic==1.6.5
amqp==5.0.6
bcrypt==3.2.0
billiard==3.6.4.0
bleach==3.3.1
blinker==1.4
celery==5.0.5
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.0.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
prompt-toolkit==3.0.18
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.17
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
